Cleaned up some of the server code, since it was becoming unmaintainable. Added default branch creation when a new conversation is created
