+++ server/socket_io_connector.js
@@ -1,6 +1,16 @@
-import { User, Conversation, Branch, Message } from './mongo_connector';
-const connect = (socketServer) => {
-    const connections = [];
+import io from 'socket.io';
+import { saveMessage } from './db/mongo_connector';
+
+const emitter = (connections) => (messageResponse) => {
+    connections.forEach( connectedSocket => {
+        connectedSocket.emit('message', messageResponse);
+    });
+}
+
+
+export const connectIO = (server) => {
+    let socketServer = io(server)
+    let connections = [];
 
     socketServer.on('connection', socket => {
         console.log('connected');
@@ -8,43 +18,7 @@ const connect = (socketServer) => {
 
         socket.on('message', action => {
             console.log('MESSAGE')
-            let newMessage = new Message({ 
-                body: action.message,
-            })
-            User.findOne({ name: action.sender.name })
-                .exec((err, user) => {
-                    console.log(user.name)
-                    newMessage.sender = user._id;
-                    Conversation.findOne({ name: action.sender.currentConversation })
-                        .populate({
-                            path: 'branches',
-                            matches: { name: action.sender.currentBranch },
-                            options: { limit: 1 }
-                        })
-                        .exec((err, conversation) => {
-                            let branch = conversation.branches[0];
-                            console.log(branch.name)
-                            branch.messages.push(newMessage._id);
-                            branch.markModified('messages');
-                            branch.save();
-                            newMessage.branch = branch._id;
-                            newMessage.save();
-
-                            let messageResponse = {
-                                body: newMessage.body,
-                                sender: { name: user.name },
-                                time: newMessage.time,
-                            }
-                            console.log(messageResponse);
-                            connections.forEach( connectedSocket => {
-                                //if (connectedSocket !== socket) {
-                                console.log('emitting message', action)
-                                connectedSocket.emit('message', messageResponse);
-                                //}
-                            });
-                        })
-                    
-                })
+            saveMessage(action, emitter(connections))
         });
 
         socket.on('disconnect', () => {
@@ -54,5 +28,3 @@ const connect = (socketServer) => {
         });
     });
 }
-
-export default connect;
