+++ src/actions/actions.js
@@ -1,20 +1,21 @@
 import ActionTypes from './actionTypes';
+import { Defaults } from '../utils/defaults';
 
 export const requestLoadConversation = (conversationId) => ({
     type: ActionTypes.LOAD_CONVERSATION_REQUEST,
     id: conversationId,
-    defaultBranchName: "DEFAULT",
+    defaultBranchName: Defaults.defaultBranchName,
 });
 export const failLoadConversation = (conversationId, error) => ({
     type: ActionTypes.LOAD_CONVERSATION_FAIL,
     id: conversationId,
-    defaultBranchName: "DEFAULT",
+    defaultBranchName: Defaults.defaultBranchName,
     error: error
 });
 export const succeedLoadConversation = (conversationId) => ({
     type: ActionTypes.LOAD_CONVERSATION_SUCCESS,
     id: conversationId,
-    defaultBranchName: "DEFAULT",
+    defaultBranchName: Defaults.defaultBranchName,
 });
 
 export const changeMessageText = (newText) => ({
@@ -126,18 +127,18 @@ export const succeedDeleteBranch = (branchName) => ({
 export const requestAddConversation = (conversationName) => ({
     type: ActionTypes.ADD_CONVERSATION_REQUEST,
     conversationName: conversationName,
-    defaultBranchName: "DEFAULT",
+    defaultBranchName: Defaults.defaultBranchName,
 });
 export const failAddConversation = (conversationName, error) => ({
     type: ActionTypes.ADD_CONVERSATION_FAIL,
     conversationName: conversationName,
-    defaultBranchName: "DEFAULT",
+    defaultBranchName: Defaults.defaultBranchName,
     error: error,
 });
 export const succeedAddConversation = (conversationName) => ({
     type: ActionTypes.ADD_CONVERSATION_SUCCESS,
     conversationName: conversationName,
-    defaultBranchName: "DEFAULT",
+    defaultBranchName: Defaults.defaultBranchName,
 });
 
 export const requestCheckoutConversation = (conversationName) => ({
