+++ src/containers/VisibleConversationPreviewList.js
@@ -2,6 +2,7 @@ import { connect } from 'react-redux';
 import { loadAllMessages, loadAllConversations } from '../actions/asyncActions';
 import ConversationPreviewList from '../components/ConversationPreviewList';
 import { push } from 'react-router-redux';
+import { Defaults } from '../utils/defaults';
 
 const getVisibleConversation = (userInfo, conversations) => {
     return conversations.allIds;
@@ -13,10 +14,9 @@ const mapStateToProps = (state) => ({
 });
 
 const mapDispatchToProps = (dispatch) => ({
-    onComponentMounted: (currentUser) => dispatch(loadAllConversations(currentUser.name)),
+    onComponentMounted: (currentUser) => dispatch(loadAllConversations(currentUser.name, Defaults.defaultBranchName)),
     onConversationClicked: (currentUser, conversationId) => {
-        const branchId = 'DEFAULT';
-        dispatch(loadAllMessages(currentUser, conversationId, branchId));
+        dispatch(loadAllMessages(currentUser, conversationId, Defaults.defaultBranchName));
         dispatch(push(`/${conversationId}`));
     }
 });
