+++ src/utils/defaults.js
@@ -1,1 +1,4 @@
 //export const SERVER_URL = 'http://localhost:3000';
+export const Defaults = {
+    defaultBranchName: 'DEFAULT', 
+}
